Make `reset` the default
Closes GH-81.

This adds proper support for presets in **remark-lint**.

This is a big one: **remark-lint** no longer does anything
without you telling it to.

* Update internals, docs, examples, to turn `reset`
  on by default;
* Refactor project to monorepo;
* Move library to `./packages/remark-lint`;
* Add the following preset packages:
  * `remark-preset-lint-consistent`;
  * `remark-preset-lint-recommended`;
  * `remark-preset-lint-legacy`.

If you’re using the CLI, please add one or more of these presets
to your configuration file.

On the API, note that presets are node modules, so it’s possible
to `require` them as well.