Rewrite `table-cell-padding` with several fixes
Basically, the rule is rewritten as the previous one did not work
as expected.  The behaviour now is what previously was intended.
More cells are checked.  When compact, everything but the biggest
cell is ignored (as shorter cells may be padded for aligning fences).
Empty body- and heading cells are also handled properly.

Finally, there’s a better message when `'padded'`, if the biggest
cell is padded more than required.

Closes GH-108.